Swithc grpc client based on browser enviroment
